***** This fixes a compiler problem ***** ***** This fixes a compiler problem ***** ***** This fixes a compiler problem *****
* Take tag {@code development-24} from the respository. Compile with eclipse
   (v3.2.1) via Project/Clean... Run all unit tests. All tests pass.
* Turn to your terminal, and run the tests with Maven 2 (mvn test).
   Nothing is compiled (the class files generated by eclipse are recent enough),
   and all tests pass.
* Clean, compile and test with Maven 2 (on Mac OS X with all (Java) updates
   applied; java -version says "build 1.5.0_07-87"): do
   mvn clean test. You get 25 errors!
* Turn to eclipse, and run all unit tests there again: you get 25 errors!
   n inspection, the error turns out to be a AbstractMethodError.

**There is obviously a difference in the byte code generated by the 2 compilers.**

What should happen is that an implementation for this method is injected
from AbstractCollectionEvent, although the method is redefined in
SetEvent with a stronger return type. It "magically matches".

This does work with the eclipse compiler. The compiler used by Maven
doesn't do this correctly, but there are no compilation errors.
Obviously, it is that compiler that is at fault.

The workaround is to explicitly do the injection, what is what we do here.

It is unclear whether this is a problem in javac in general, or only
in this version, or in the Mac OS X version.

The problem was observed with ActualSetEvent. The fix is applied on all
actual events of that hierarchy.

The fix should be removed if the problem in the compiler is solved. Also,
the methods in the super classes should revert to final.