Halleluja! This is the result of a multi-day effort, simplifying starting the topological update ("update dependents') from different points. All tests GREEN, but tests in general should maybe be revisited (or new test set created from scratch, or Toryt, ...). The change is big, but should be all-internal, and no user code should be influenced.
All "packageUpdateDependents" methods are removed. The call to "updateDependents" in Edit now just calls the topological update itself.
It uses events created by the edit. For this "createEvent" was changed to "createEvents", so the algorithm can service both simple edits, complex exits, and compound edits.

Helper methods were added to AbstractDependentBeed and AbstractEdit to help with this. AbstractEdit now features a method to create a singleton source event map for topological update from 1 event (to ease implementation of createEvents for simple cases). AbstractDependentBeed features an updateDependents(Event) method, for use with structural changes (non-edit changes). Other "updateDependents" overloadings are removed.

This makes AbstractEdit very empty. This class was intended to protect Beedraz users from exposure to needles generics (the _Event_ type of Edits). Now, there is almost nothing left, and maybe we should simply remove this class.

Also, we now know more or less for sure we need AbstractBeed for topological update. This is needed because the topological update needs access, privileged, to "fireEvent", for sending events to regular MVC listeners. So, all actual beeds must inherit from AbstractBeed. The interface Beed is still needed for multiple inheritance. Now, abstract edits also demand that the _Target_ extends AbstractBeed, instead of only beed. This is needed for the new updateDependents code. In the above context, this makes sense.

Furthermore, we employ less strict generics in the map of source events passed to the topological update method (? extends Event).

Getting the tests work was a bit of a dare, since the design is now "better" in the respect that more methods are shielded from use by external users, and have more (asserted) preconditions. We needed some workarounds to get the tests working again. A pattern is becoming clear, what shows we maybe should rework tests from the top down.

Also, some obsolete MUDO comments were removed.