AggregateEvent.getComponentevents() renamed to AggregateEvent.getComponentEvents().
For CompoundEdit, we need to be able to combine consecutive events from a common source. The necessary method is introduced in Event, and implemented in all events. This is easy, except for ActualMapEvent: this is very, very difficult.

All test are GREEN, but no test have been created for this new code yet.

In concept, creating and performing a compound edit should now work.